@@ -1,6 +1,7 @@
 #
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
+
 from __future__ import annotations
 
 import datetime
@@ -10,7 +11,6 @@ import os
 import subprocess
 import uuid
 from pathlib import Path
-from typing import Optional
 
 import anyio
 import asyncer
@@ -32,7 +32,7 @@ class ConnectorRunner:
         dagger_client: dagger.Client,
         execution_inputs: ExecutionInputs,
         is_airbyte_ci: bool,
-        http_proxy: Optional[Proxy] = None,
+        http_proxy: Proxy | None = None,
     ):
         self.connector_under_test = execution_inputs.connector_under_test
         self.command = execution_inputs.command
@@ -49,6 +49,9 @@ class ConnectorRunner:
         self.http_proxy = http_proxy
         self.logger = logging.getLogger(f"{self.connector_under_test.name}-{self.connector_under_test.version}")
         self.dagger_client = dagger_client.pipeline(f"{self.connector_under_test.name}-{self.connector_under_test.version}")
+
+        # When invoked via airbyte-ci, record_obfuscator.py is copied over to /tmp
+        # but when running locally, it's in Airbyte repo root.
         if is_airbyte_ci:
             self.host_obfuscator_path = "/tmp/record_obfuscator.py"
         else:
@@ -68,6 +71,7 @@ class ConnectorRunner:
         return (self.output_dir / "stderr.log").resolve()
 
     def _get_full_command(self, command: Command) -> list[str]:
+        """Returns a list with a full Airbyte command invocation and all it's arguments and options."""
         if command is Command.SPEC:
             return ["spec"]
         elif command is Command.CHECK:
@@ -95,10 +99,10 @@ class ConnectorRunner:
         else:
             raise NotImplementedError(f"The connector runner does not support the {command} command")
 
-    async def get_container_env_variable_value(self, name: str) -> Optional[str]:
+    async def get_container_env_variable_value(self, name: str) -> str | None:
         return await self._connector_under_test_container.env_variable(name)
 
-    async def get_container_label(self, label: str) -> Optional[str]:
+    async def get_container_label(self, label: str) -> str | None:
         return await self._connector_under_test_container.label(label)
 
     async def get_container_entrypoint(self) -> str:
@@ -118,12 +122,14 @@ class ConnectorRunner:
         container = self._connector_under_test_container
         # Do not cache downstream dagger layers
         container = container.with_env_variable("CACHEBUSTER", str(uuid.uuid4()))
-        expanded_host_executable_path = os.path.expanduser(self.host_obfuscator_path)
 
+        # When running locally, it's likely that record_obfuscator is within the user's home directory, so we expand it.
+        expanded_host_executable_path = os.path.expanduser(self.host_obfuscator_path)
         container = container.with_file(
             self.IN_CONTAINER_OBFUSCATOR_PATH,
             self.dagger_client.host().file(expanded_host_executable_path),
         )
+
         for env_var_name, env_var_value in self.environment_variables.items():
             container = container.with_env_variable(env_var_name, env_var_value)
         if self.config:
@@ -144,7 +150,7 @@ class ConnectorRunner:
             entrypoint = await container.entrypoint()
             assert entrypoint, "The connector container has no entrypoint"
             airbyte_command = entrypoint + self.full_command
-            # We are piping the output to a file to avoidQueryError: file size exceeds limit 134217728
+            # We are piping the output to a file to avoid QueryError: file size exceeds limit 134217728
             container = container.with_exec(
                 [
                     "sh",
@@ -174,6 +180,7 @@ class ConnectorRunner:
             self.logger.error(f"❌ Failed to run {self.command.value} command")
         else:
             self.logger.info(f"⌛ Finished running {self.command.value} command")
+
         execution_result = await ExecutionResult.load(
             command=self.command,
             connector_under_test=self.connector_under_test,
