@@ -4,9 +4,10 @@
 
 package io.airbyte.cdk.load.task.implementor
 
+import io.airbyte.cdk.load.state.CheckpointManager
 import io.airbyte.cdk.load.state.SyncManager
 import io.airbyte.cdk.load.task.DestinationTaskExceptionHandler
-import io.airbyte.cdk.load.task.ImplementorTask
+import io.airbyte.cdk.load.task.ShutdownScope
 import io.airbyte.cdk.load.util.setOnce
 import io.airbyte.cdk.load.write.DestinationWriter
 import io.github.oshai.kotlinlogging.KotlinLogging
@@ -14,36 +15,43 @@ import io.micronaut.context.annotation.Secondary
 import jakarta.inject.Singleton
 import java.util.concurrent.atomic.AtomicBoolean
 
-interface FailSyncTask : ImplementorTask
+interface FailSyncTask : ShutdownScope
 
 /**
  * FailSyncTask is a task that is executed when a sync fails. It is responsible for cleaning up
  * resources and reporting the failure.
  */
 class DefaultFailSyncTask(
-    private val exceptionHandler: DestinationTaskExceptionHandler<*>,
+    private val exceptionHandler: DestinationTaskExceptionHandler<*, *>,
     private val destinationWriter: DestinationWriter,
     private val exception: Exception,
     private val syncManager: SyncManager,
+    private val checkpointManager: CheckpointManager<*, *>
 ) : FailSyncTask {
     private val log = KotlinLogging.logger {}
-    private val teardownRan = AtomicBoolean(false)
+
+    companion object {
+        private val syncFailedHasRun = AtomicBoolean(false)
+    }
 
     override suspend fun execute() {
-        if (teardownRan.setOnce()) {
+        if (syncFailedHasRun.setOnce()) {
+            // Ensure any remaining ready state gets captured: don't waste work!
+            checkpointManager.flushReadyCheckpointMessages()
             val result = syncManager.markFailed(exception) // awaits stream completion
             log.info { "Calling teardown with failure result $result" }
+            exceptionHandler.handleSyncFailed()
+            // Do this cleanup last, after all the tasks have had a decent chance to finish.
             destinationWriter.teardown(result)
-            exceptionHandler.handleTeardownComplete()
         } else {
-            log.info { "Teardown already ran, doing nothing." }
+            log.info { "Fail sync task already initiated, doing nothing." }
         }
     }
 }
 
 interface FailSyncTaskFactory {
     fun make(
-        exceptionHandler: DestinationTaskExceptionHandler<*>,
+        exceptionHandler: DestinationTaskExceptionHandler<*, *>,
         exception: Exception
     ): FailSyncTask
 }
@@ -52,12 +60,20 @@ interface FailSyncTaskFactory {
 @Secondary
 class DefaultFailSyncTaskFactory(
     private val syncManager: SyncManager,
+    private val checkpointManager: CheckpointManager<*, *>,
     private val destinationWriter: DestinationWriter
 ) : FailSyncTaskFactory {
+
     override fun make(
-        exceptionHandler: DestinationTaskExceptionHandler<*>,
+        exceptionHandler: DestinationTaskExceptionHandler<*, *>,
         exception: Exception
     ): FailSyncTask {
-        return DefaultFailSyncTask(exceptionHandler, destinationWriter, exception, syncManager)
+        return DefaultFailSyncTask(
+            exceptionHandler,
+            destinationWriter,
+            exception,
+            syncManager,
+            checkpointManager
+        )
     }
 }
