@@ -5,13 +5,13 @@
 package io.airbyte.cdk.load.task.implementor
 
 import io.airbyte.cdk.load.task.DestinationTaskLauncher
-import io.airbyte.cdk.load.task.ImplementorTask
-import io.airbyte.cdk.load.task.SyncTask
+import io.airbyte.cdk.load.task.ImplementorScope
+import io.airbyte.cdk.load.task.SyncLevel
 import io.airbyte.cdk.load.write.DestinationWriter
 import io.micronaut.context.annotation.Secondary
 import jakarta.inject.Singleton
 
-interface SetupTask : SyncTask, ImplementorTask
+interface SetupTask : SyncLevel, ImplementorScope
 
 /**
  * Wraps @[DestinationWriter.setup] and starts the open stream tasks.
