@@ -6,6 +6,7 @@ from contextlib import nullcontext as does_not_raise
 from pathlib import Path
 
 import pytest
+import semver
 from connector_ops import utils
 
 
@@ -45,7 +46,7 @@ class TestConnector:
             assert isinstance(connector.support_level, str)
             assert isinstance(connector.acceptance_test_config, dict)
             assert connector.icon_path == Path(f"./airbyte-integrations/connectors/{connector.technical_name}/icon.svg")
-            assert len(connector.version.split(".")) == 3
+            assert semver.Version.parse(connector.version)
         else:
             assert connector.metadata is None
             assert connector.support_level is None
