@@ -316,7 +316,7 @@ def prompt_for_read_with_or_without_state() -> bool:
     📖 Do you want to run the read command with or without state?
     1. Run the read command with state
     2. Run the read command without state
-                              
+
     We recommend reading with state to properly test incremental sync.
     But if the target version introduces a breaking change in the state, you might want to run without state.
     """
