@@ -6,28 +6,30 @@ import json
 import os
 import uuid
 from datetime import datetime
-from typing import Dict, List, Tuple
+from pathlib import Path
+from typing import Dict, Iterable, List, Tuple
 
 import anyio
+import semver
+import yaml
 from airbyte_protocol.models.airbyte_protocol import ConnectorSpecification  # type: ignore
-from connector_ops.utils import ConnectorLanguage  # type: ignore
-from dagger import Container, ExecError, File, ImageLayerCompression, Platform, QueryError
+from connector_ops.utils import METADATA_FILE_NAME, ConnectorLanguage  # type: ignore
+from dagger import Container, Directory, ExecError, File, ImageLayerCompression, Platform, QueryError
 from pipelines import consts
 from pipelines.airbyte_ci.connectors.build_image import steps
 from pipelines.airbyte_ci.connectors.publish.context import PublishConnectorContext, RolloutMode
 from pipelines.airbyte_ci.connectors.reports import ConnectorReport
-from pipelines.airbyte_ci.metadata.pipeline import (
-    MetadataPromoteReleaseCandidate,
-    MetadataRollbackReleaseCandidate,
-    MetadataUpload,
-    MetadataValidation,
-)
+from pipelines.airbyte_ci.metadata.pipeline import MetadataRollbackReleaseCandidate, MetadataUpload, MetadataValidation
+from pipelines.airbyte_ci.steps.bump_version import SetConnectorVersion
+from pipelines.airbyte_ci.steps.changelog import AddChangelogEntry
+from pipelines.airbyte_ci.steps.pull_request import CreateOrUpdatePullRequest
 from pipelines.airbyte_ci.steps.python_registry import PublishToPythonRegistry, PythonRegistryPublishContext
 from pipelines.consts import LOCAL_BUILD_PLATFORM
 from pipelines.dagger.actions.remote_storage import upload_to_gcs
 from pipelines.dagger.actions.system import docker
+from pipelines.helpers.connectors.dagger_fs import dagger_read_file, dagger_write_file
 from pipelines.helpers.pip import is_package_published
-from pipelines.models.steps import Step, StepResult, StepStatus
+from pipelines.models.steps import Step, StepModifyingFiles, StepResult, StepStatus
 from pydantic import BaseModel, ValidationError
 
 
@@ -384,6 +386,74 @@ class UploadSbom(Step):
         return StepResult(step=self, status=StepStatus.SUCCESS, stdout="Uploaded SBOM to metadata service bucket.")
 
 
+class SetPromotedVersion(SetConnectorVersion):
+    context: PublishConnectorContext
+    title = "Promote release candidate"
+
+    @property
+    def current_semver_version(self) -> semver.Version:
+        return semver.Version.parse(self.context.connector.version)
+
+    @property
+    def promoted_semver_version(self) -> semver.Version:
+        return self.current_semver_version.replace(prerelease=None)
+
+    @property
+    def promoted_version(self) -> str:
+        return str(self.promoted_semver_version)
+
+    @property
+    def current_version_is_rc(self) -> bool:
+        return bool(self.current_semver_version.prerelease and "rc" in self.current_semver_version.prerelease)
+
+    def __init__(self, context: PublishConnectorContext, connector_directory: Directory) -> None:
+        self.context = context
+        super().__init__(context, connector_directory, self.promoted_version)
+
+    async def _run(self) -> StepResult:
+        if not self.current_version_is_rc:
+            return StepResult(step=self, status=StepStatus.SKIPPED, stdout="The connector version has no rc suffix.")
+        return await super()._run()
+
+
+class ResetReleaseCandidate(StepModifyingFiles):
+    context: PublishConnectorContext
+    title = "Reset release candidate flag"
+
+    async def _run(self) -> StepResult:
+        raw_metadata = await dagger_read_file(await self.context.get_connector_dir(include=METADATA_FILE_NAME), METADATA_FILE_NAME)
+        current_metadata = yaml.safe_load(raw_metadata)
+        is_release_candidate = current_metadata.get("data", {}).get("releases", {}).get("isReleaseCandidate", False)
+        if not is_release_candidate:
+            return StepResult(step=self, status=StepStatus.SKIPPED, stdout="The connector is not a release candidate.")
+        # We do an in-place replacement instead of serializing back to yaml to preserve comments and formatting.
+        new_raw_metadata = raw_metadata.replace("isReleaseCandidate: true", "isReleaseCandidate: false")
+        self.modified_directory = dagger_write_file(self.modified_directory, METADATA_FILE_NAME, new_raw_metadata)
+        self.modified_files.append(METADATA_FILE_NAME)
+        return StepResult(
+            step=self,
+            status=StepStatus.SUCCESS,
+            stdout="Set the isReleaseCandidate flag to false in the metadata file.",
+            output=self.modified_directory,
+        )
+
+
+# Helpers
+def create_connector_report(results: List[StepResult], context: PublishConnectorContext) -> ConnectorReport:
+    """Generate a connector report from results and assign it to the context.
+
+    Args:
+        results (List[StepResult]): List of step results.
+        context (PublishConnectorContext): The connector context to assign the report to.
+
+    Returns:
+        ConnectorReport: The connector report.
+    """
+    report = ConnectorReport(context, results, name="PUBLISH RESULTS")
+    context.report = report
+    return report
+
+
 # Pipeline
 async def run_connector_publish_pipeline(context: PublishConnectorContext, semaphore: anyio.Semaphore) -> ConnectorReport:
     """Run a publish pipeline for a single connector.
@@ -415,11 +485,6 @@ async def run_connector_publish_pipeline(context: PublishConnectorContext, semap
 
     upload_sbom_step = UploadSbom(context)
 
-    def create_connector_report(results: List[StepResult]) -> ConnectorReport:
-        report = ConnectorReport(context, results, name="PUBLISH RESULTS")
-        context.report = report
-        return report
-
     async with semaphore:
         async with context:
             results = []
@@ -429,14 +494,14 @@ async def run_connector_publish_pipeline(context: PublishConnectorContext, semap
 
             # Exit early if the metadata file is invalid.
             if metadata_validation_results.status is not StepStatus.SUCCESS:
-                return create_connector_report(results)
+                return create_connector_report(results, context)
 
             check_connector_image_results = await CheckConnectorImageDoesNotExist(context).run()
             results.append(check_connector_image_results)
             python_registry_steps, terminate_early = await _run_python_registry_publish_pipeline(context)
             results.extend(python_registry_steps)
             if terminate_early:
-                return create_connector_report(results)
+                return create_connector_report(results, context)
 
             # If the connector image already exists, we don't need to build it, but we still need to upload the metadata file.
             # We also need to upload the spec to the spec cache bucket.
@@ -448,26 +513,26 @@ async def run_connector_publish_pipeline(context: PublishConnectorContext, semap
                 upload_to_spec_cache_results = await upload_spec_to_cache_step.run(already_published_connector)
                 results.append(upload_to_spec_cache_results)
                 if upload_to_spec_cache_results.status is not StepStatus.SUCCESS:
-                    return create_connector_report(results)
+                    return create_connector_report(results, context)
 
                 upload_sbom_results = await upload_sbom_step.run()
                 results.append(upload_sbom_results)
                 if upload_sbom_results.status is not StepStatus.SUCCESS:
-                    return create_connector_report(results)
+                    return create_connector_report(results, context)
 
                 metadata_upload_results = await metadata_upload_step.run()
                 results.append(metadata_upload_results)
 
             # Exit early if the connector image already exists or has failed to build
             if check_connector_image_results.status is not StepStatus.SUCCESS:
-                return create_connector_report(results)
+                return create_connector_report(results, context)
 
             build_connector_results = await steps.run_connector_build(context)
             results.append(build_connector_results)
 
             # Exit early if the connector image failed to build
             if build_connector_results.status is not StepStatus.SUCCESS:
-                return create_connector_report(results)
+                return create_connector_report(results, context)
 
             if context.connector.language in [ConnectorLanguage.PYTHON, ConnectorLanguage.LOW_CODE]:
                 upload_dependencies_step = await UploadDependenciesToMetadataService(context).run(build_connector_results.output)
@@ -479,7 +544,7 @@ async def run_connector_publish_pipeline(context: PublishConnectorContext, semap
 
             # Exit early if the connector image failed to push
             if push_connector_image_results.status is not StepStatus.SUCCESS:
-                return create_connector_report(results)
+                return create_connector_report(results, context)
 
             # Make sure the image published is healthy by pulling it and running SPEC on it.
             # See https://github.com/airbytehq/airbyte/issues/26085
@@ -488,21 +553,21 @@ async def run_connector_publish_pipeline(context: PublishConnectorContext, semap
 
             # Exit early if the connector image failed to pull
             if pull_connector_image_results.status is not StepStatus.SUCCESS:
-                return create_connector_report(results)
+                return create_connector_report(results, context)
 
             upload_to_spec_cache_results = await upload_spec_to_cache_step.run(built_connector_platform_variants[0])
             results.append(upload_to_spec_cache_results)
             if upload_to_spec_cache_results.status is not StepStatus.SUCCESS:
-                return create_connector_report(results)
+                return create_connector_report(results, context)
 
             upload_sbom_results = await upload_sbom_step.run()
             results.append(upload_sbom_results)
             if upload_sbom_results.status is not StepStatus.SUCCESS:
-                return create_connector_report(results)
+                return create_connector_report(results, context)
 
             metadata_upload_results = await metadata_upload_step.run()
             results.append(metadata_upload_results)
-            connector_report = create_connector_report(results)
+            connector_report = create_connector_report(results, context)
     return connector_report
 
 
@@ -565,8 +630,24 @@ async def run_connector_rollback_pipeline(context: PublishConnectorContext, sema
                     context, context.metadata_bucket_name, context.metadata_service_gcs_credentials
                 ).run()
             )
+            connector_report = create_connector_report(results, context)
 
-    return ConnectorReport(context, results, name="ROLLBACK RESULTS")
+    return connector_report
+
+
+def get_promotion_pr_creation_arguments(
+    modified_files: Iterable[Path],
+    context: PublishConnectorContext,
+    step_results: Iterable[StepResult],
+    release_candidate_version: str,
+    promoted_version: str,
+) -> Tuple[Tuple, Dict]:
+    return (modified_files,), {
+        "branch_id": f"{context.connector.technical_name}/{promoted_version}",
+        "commit_message": "\n".join(step_result.step.title for step_result in step_results if step_result.success),
+        "pr_title": f"🐙 {context.connector.technical_name}: release {promoted_version}",
+        "pr_body": f"The release candidate version {release_candidate_version} has been deemed stable and is now ready to be promoted to an official release ({promoted_version}).",
+    }
 
 
 async def run_connector_promote_pipeline(context: PublishConnectorContext, semaphore: anyio.Semaphore) -> ConnectorReport:
@@ -578,22 +659,69 @@ async def run_connector_promote_pipeline(context: PublishConnectorContext, semap
     Returns:
         ConnectorReport: The reports holding promote results.
     """
+
     results = []
+    current_version = context.connector.version
+    all_modified_files = set()
     async with semaphore:
         async with context:
             assert context.rollout_mode == RolloutMode.PROMOTE, "This pipeline can only run in promote mode."
-            assert context.connector.metadata.get("releases", {}).get(
-                "isReleaseCandidate", True
-            ), "This pipeline can only run for release candidates."
-            metadata_promote_result = await MetadataPromoteReleaseCandidate(
-                context, context.metadata_bucket_name, context.metadata_service_gcs_credentials
-            ).run()
-            results.append(metadata_promote_result)
-            if metadata_promote_result.status is StepStatus.FAILURE:
-                return ConnectorReport(context, results, name="PROMOTE RESULTS")
-            publish_latest_tag_results = await PushVersionImageAsLatest(context).run()
-            results.append(publish_latest_tag_results)
-    return ConnectorReport(context, results, name="PROMOTE RESULTS")
+            original_connector_directory = await context.get_connector_dir()
+            # Remove RC suffix
+            set_promoted_version = SetPromotedVersion(context, original_connector_directory)
+            set_promoted_version_results = await set_promoted_version.run()
+            results.append(set_promoted_version_results)
+            if set_promoted_version_results.success:
+                all_modified_files.update(await set_promoted_version.export_modified_files(context.connector.code_directory))
+
+            # Set isReleaseCandidate to False
+            reset_release_candidate = ResetReleaseCandidate(context, set_promoted_version_results.output)
+            reset_release_candidate_results = await reset_release_candidate.run()
+            results.append(reset_release_candidate_results)
+            if reset_release_candidate_results.success:
+                all_modified_files.update(await reset_release_candidate.export_modified_files(context.connector.code_directory))
+
+            if not all([result.success for result in results]):
+                context.logger.error("The metadata update failed. Skipping PR creation.")
+                connector_report = create_connector_report(results, context)
+                return connector_report
+
+            # Open PR when all previous steps are successful
+            promoted_version = set_promoted_version.promoted_version
+            initial_pr_creation = CreateOrUpdatePullRequest(context, skip_ci=True)
+            pr_creation_args, pr_creation_kwargs = get_promotion_pr_creation_arguments(
+                all_modified_files, context, results, current_version, promoted_version
+            )
+            initial_pr_creation_result = await initial_pr_creation.run(*pr_creation_args, **pr_creation_kwargs)
+            results.append(initial_pr_creation_result)
+            # Update changelog and update PR
+            if initial_pr_creation_result.success:
+                created_pr = initial_pr_creation_result.output
+                documentation_directory = await context.get_repo_dir(
+                    include=[str(context.connector.local_connector_documentation_directory)]
+                ).directory(str(context.connector.local_connector_documentation_directory))
+                add_changelog_entry = AddChangelogEntry(
+                    context,
+                    documentation_directory,
+                    promoted_version,
+                    f"Promoting release candidate {current_version} to a main version.",
+                    created_pr.number,
+                )
+                add_changelog_entry_result = await add_changelog_entry.run()
+                results.append(add_changelog_entry_result)
+                if add_changelog_entry_result.success:
+                    all_modified_files.update(
+                        await add_changelog_entry.export_modified_files(context.connector.local_connector_documentation_directory)
+                    )
+                post_changelog_pr_update = CreateOrUpdatePullRequest(context, skip_ci=False, labels=["auto-merge"])
+                pr_creation_args, pr_creation_kwargs = get_promotion_pr_creation_arguments(
+                    all_modified_files, context, results, current_version, promoted_version
+                )
+                post_changelog_pr_update_result = await post_changelog_pr_update.run(*pr_creation_args, **pr_creation_kwargs)
+                results.append(post_changelog_pr_update_result)
+
+            connector_report = create_connector_report(results, context)
+    return connector_report
 
 
 def reorder_contexts(contexts: List[PublishConnectorContext]) -> List[PublishConnectorContext]:
