@@ -1,9 +1,10 @@
 # Copyright (c) 2024 Airbyte, Inc., all rights reserved.
 
-import logging
 from functools import wraps
 from time import sleep
 from typing import Any, Callable, Final, Optional, Tuple, Type
+
+from source_shopify.utils import LOGGER
 
 from .exceptions import ShopifyBulkExceptions
 
@@ -13,11 +14,10 @@ BULK_RETRY_ERRORS: Final[Tuple] = (
 )
 
 
-def bulk_retry_on_exception(logger: logging.Logger, more_exceptions: Optional[Tuple[Type[Exception], ...]] = None) -> Callable:
+def bulk_retry_on_exception(more_exceptions: Optional[Tuple[Type[Exception], ...]] = None) -> Callable:
     """
     A decorator to retry a function when specified exceptions are raised.
 
-    :param logger: Number of times to retry.
     :param more_exceptions: A tuple of exception types to catch.
     """
 
@@ -31,26 +31,26 @@ def bulk_retry_on_exception(logger: logging.Logger, more_exceptions: Optional[Tu
                 except BULK_RETRY_ERRORS or more_exceptions as ex:
                     current_retries += 1
                     if current_retries > self._job_max_retries:
-                        logger.error("Exceeded retry limit. Giving up.")
+                        LOGGER.error("Exceeded retry limit. Giving up.")
                         raise
                     else:
-                        logger.warning(
+                        LOGGER.warning(
                             f"Stream `{self.http_client.name}`: {ex}. Retrying {current_retries}/{self._job_max_retries} after {self._job_backoff_time} seconds."
                         )
                         sleep(self._job_backoff_time)
                 except ShopifyBulkExceptions.BulkJobCreationFailedConcurrentError:
                     if self._concurrent_attempt == self._concurrent_max_retry:
                         message = f"The BULK Job couldn't be created at this time, since another job is running."
-                        logger.error(message)
+                        LOGGER.error(message)
                         raise ShopifyBulkExceptions.BulkJobConcurrentError(message)
 
                     self._concurrent_attempt += 1
-                    logger.warning(
+                    LOGGER.warning(
                         f"Stream: `{self.http_client.name}`, the BULK concurrency limit has reached. Waiting {self._concurrent_interval} sec before retry, attempt: {self._concurrent_attempt}.",
                     )
                     sleep(self._concurrent_interval)
                 except ShopifyBulkExceptions.BulkJobRedirectToOtherShopError:
-                    logger.warning(
+                    LOGGER.warning(
                         f"Stream: `{self.http_client.name}`, the `shop name` differs from the provided in `input configuration`. Switching to the `{self._tools.shop_name_from_url(self.base_url)}`.",
                     )
 
